requirements (1).txt
flask
twilio
anthropic
requests
numpy
opencv-python
ultralytics
face-recognition
Pillow
